+++ CHANGELOG.md
@@ -4,7 +4,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 -  nothing yet
 
-## [v2.0.0](https://github.com/regevbr/busywait.js/releases/tag/v2.0.0)
+## [v3.0.0](https://github.com/regevbr/busywait.js/compare/v2.0.0...v3.0.0)
+### Breaking changes
+-  Errors are now an instance of `Error` and not plain strings
+-  Options type was modified
+-  Return type was modified
+### Added
+-  Exponential backoff (with optional full jitter) support
+-  Added to the result, the time it took to finish
+### Fixed
+-  Better type inference
+-  Updated readme
+-  Updated all dependencies to the latest versions
+
+## [v2.0.0](https://github.com/regevbr/busywait.js/compare/v1.1.0...v2.0.0)
 ### Breaking changes
 -  Removed redundant export methods
 -  Sync check functions will be considered as failed only if they throw error
@@ -14,5 +27,3 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 -  Updated readme
 -  Updated all dependencies to the latest versions 
-
-[v2.0.0]: https://github.com/PruvoNet/squiss-ts/compare/v1.1.0...v2.0.0